Move CancelModal back to component
